// src/queries/model/responseProductSourcingDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */
import type { ResponseProductSourcingDtoAplyStts } from './responseProductSourcingDtoAplyStts';
import type { ResponseCompanyDto } from './responseCompanyDto';
import type { ResponseProductSourcingDtoStlmStts } from './responseProductSourcingDtoStlmStts';
import type { UserResponseDto } from './userResponseDto';

export interface ResponseProductSourcingDto {
  /** 신청내용 */
  aplyCn: string;
  /** 의뢰신청명 */
  aplymNm: string;
  /** 신청상태 */
  aplyStts: ResponseProductSourcingDtoAplyStts;
  /** 브랜드사 */
  company: ResponseCompanyDto;
  /** 수정 일시 */
  mdfcnDt?: string;
  /** 수정자 ID */
  mdfrId?: number;
  /** 상품 분류값 */
  prdctClsfVl: string;
  /** 등록 일시 */
  regDt: string;
  /** 등록자 ID */
  rgtrId: number;
  /** 소싱ID */
  srcngId: number;
  /** 결제ID */
  stlmId?: string;
  /** 결제상태 */
  stlmStts: ResponseProductSourcingDtoStlmStts;
  /** 유저 이메일 */
  user: UserResponseDto;
}